requirements.txt
google-cloud-vision==2.6.2
google-cloud-storage==1.43.0
ebooklib==0.17.1